requirements.txt
fpdf
python-dotenv
supabase
psycopg2
